# api/stablemanager/models/action_summary.py
# coding: utf-8

from __future__ import absolute_import
from .base_model_ import Model
from datetime import date, datetime
from typing import List, Dict
from ..util import deserialize_model


class ActionSummary(Model):
    """
    NOTE: This class is auto generated by the swagger code generator program.
    Do not edit the class manually.
    """
    def __init__(self, id: str=None, summary: str=None):
        """
        ActionSummary - a model defined in Swagger

        :param id: The id of this ActionSummary.
        :type id: str
        :param summary: The summary of this ActionSummary.
        :type summary: str
        """
        self.swagger_types = {
            'id': str,
            'summary': str
        }

        self.attribute_map = {
            'id': 'id',
            'summary': 'summary'
        }

        self._id = id
        self._summary = summary

    @classmethod
    def from_dict(cls, dikt) -> 'ActionSummary':
        """
        Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The ActionSummary of this ActionSummary.
        :rtype: ActionSummary
        """
        return deserialize_model(dikt, cls)

    @property
    def id(self) -> str:
        """
        Gets the id of this ActionSummary.

        :return: The id of this ActionSummary.
        :rtype: str
        """
        return self._id

    @id.setter
    def id(self, id: str):
        """
        Sets the id of this ActionSummary.

        :param id: The id of this ActionSummary.
        :type id: str
        """

        self._id = id

    @property
    def summary(self) -> str:
        """
        Gets the summary of this ActionSummary.

        :return: The summary of this ActionSummary.
        :rtype: str
        """
        return self._summary

    @summary.setter
    def summary(self, summary: str):
        """
        Sets the summary of this ActionSummary.

        :param summary: The summary of this ActionSummary.
        :type summary: str
        """

        self._summary = summary