# api/stablemanager/models/schedule_summary.py
# coding: utf-8

from __future__ import absolute_import
from .base_model_ import Model
from datetime import date, datetime
from typing import List, Dict
from ..util import deserialize_model


class ScheduleSummary(Model):
    """
    NOTE: This class is auto generated by the swagger code generator program.
    Do not edit the class manually.
    """
    def __init__(self, id: str=None, summary: str=None):
        """
        ScheduleSummary - a model defined in Swagger

        :param id: The id of this ScheduleSummary.
        :type id: str
        :param summary: The summary of this ScheduleSummary.
        :type summary: str
        """
        self.swagger_types = {
            'id': str,
            'summary': str
        }

        self.attribute_map = {
            'id': 'id',
            'summary': 'summary'
        }

        self._id = id
        self._summary = summary

    @classmethod
    def from_dict(cls, dikt) -> 'ScheduleSummary':
        """
        Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The ScheduleSummary of this ScheduleSummary.
        :rtype: ScheduleSummary
        """
        return deserialize_model(dikt, cls)

    @property
    def id(self) -> str:
        """
        Gets the id of this ScheduleSummary.

        :return: The id of this ScheduleSummary.
        :rtype: str
        """
        return self._id

    @id.setter
    def id(self, id: str):
        """
        Sets the id of this ScheduleSummary.

        :param id: The id of this ScheduleSummary.
        :type id: str
        """

        self._id = id

    @property
    def summary(self) -> str:
        """
        Gets the summary of this ScheduleSummary.

        :return: The summary of this ScheduleSummary.
        :rtype: str
        """
        return self._summary

    @summary.setter
    def summary(self, summary: str):
        """
        Sets the summary of this ScheduleSummary.

        :param summary: The summary of this ScheduleSummary.
        :type summary: str
        """

        self._summary = summary